refactor: Renamed registries
