+++ hats/plugins/item_models.py
@@ -4,7 +4,7 @@ from logging import getLogger
 from beet import Context, Model, ResourcePack
 
 from hats.options import HatsOptions
-from hats.registry.hats import HatRegistry
+from hats.registries.hats import HatRegistry
 
 logger = getLogger(__name__)
