+++ hats/registries/hats.py
