+++ hats/registries/settings.py
